blip_diffusion_model

+++ lavis/models/clip_models/model.py
@@ -765,9 +765,9 @@ def convert_weights_to_fp16(model: nn.Module):
 
     def _convert_weights_to_fp16(l):
         if isinstance(l, (nn.Conv1d, nn.Conv2d, nn.Linear)):
-            l.weight.data = l.weight.data
+            l.weight.data = l.weight.data.half()
             if l.bias is not None:
-                l.bias.data = l.bias.data
+                l.bias.data = l.bias.data.half()
 
         if isinstance(l, nn.MultiheadAttention):
             for attr in [
@@ -778,13 +778,13 @@ def convert_weights_to_fp16(model: nn.Module):
             ]:
                 tensor = getattr(l, attr)
                 if tensor is not None:
-                    tensor.data = tensor.data
+                    tensor.data = tensor.data.half()
 
         for name in ["text_projection", "proj"]:
             if hasattr(l, name):
                 attr = getattr(l, name)
                 if attr is not None:
-                    attr.data = attr.data
+                    attr.data = attr.data.half()
 
     model.apply(_convert_weights_to_fp16)
